Rename io_manager to input_manager
